+++ src/routes/AdManage/adManage.module.scss
@@ -1,5 +1,12 @@
 @use '/src/styles/constants/colors';
 
+.pageHeader {
+  margin-left: 70px;
+  font-size: 24px;
+  font-weight: 500;
+  letter-spacing: 1px;
+}
+
 .makeAdButton {
   width: 100px;
   height: 40px;
